# crm-it-shkola-rtk/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/catalogs/vendors": {
      "get": {
        "operationId": "CatalogsController_getVendors",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "headers": {
              "X-Total-Count": {
                "description": "Общее количество записей без учёта пагинации",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VendorDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Список вендоров каталога (с пагинацией)",
        "tags": [
          "catalogs"
        ]
      },
      "post": {
        "operationId": "CatalogsController_createVendor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVendorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorDto"
                }
              }
            }
          }
        },
        "summary": "Создать вендора",
        "tags": [
          "catalogs"
        ]
      }
    },
    "/catalogs/vendors/{id}": {
      "get": {
        "operationId": "CatalogsController_getVendorById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorDto"
                }
              }
            }
          }
        },
        "summary": "Вендор по идентификатору",
        "tags": [
          "catalogs"
        ]
      },
      "patch": {
        "operationId": "CatalogsController_patchUpdateVendor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVendorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorDto"
                }
              }
            }
          }
        },
        "summary": "Обновить вендора",
        "tags": [
          "catalogs"
        ]
      },
      "put": {
        "deprecated": true,
        "operationId": "CatalogsController_updateVendor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVendorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorDto"
                }
              }
            }
          }
        },
        "summary": "Обновить вендора",
        "tags": [
          "catalogs"
        ]
      },
      "delete": {
        "operationId": "CatalogsController_deleteVendor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Удалить вендора",
        "tags": [
          "catalogs"
        ]
      }
    },
    "/catalogs/it-directions": {
      "get": {
        "operationId": "CatalogsController_getItDirections",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "headers": {
              "X-Total-Count": {
                "description": "Общее количество записей без учёта пагинации",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItDirectionDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Список ИТ-направлений (с пагинацией)",
        "tags": [
          "catalogs"
        ]
      },
      "post": {
        "operationId": "CatalogsController_createItDirection",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItDirectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItDirectionDto"
                }
              }
            }
          }
        },
        "summary": "Создать ИТ-направление",
        "tags": [
          "catalogs"
        ]
      }
    },
    "/catalogs/it-directions/{id}": {
      "patch": {
        "operationId": "CatalogsController_patchUpdateItDirection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItDirectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItDirectionDto"
                }
              }
            }
          }
        },
        "summary": "Обновить ИТ-направление",
        "tags": [
          "catalogs"
        ]
      },
      "put": {
        "deprecated": true,
        "operationId": "CatalogsController_updateItDirection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItDirectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItDirectionDto"
                }
              }
            }
          }
        },
        "summary": "Обновить ИТ-направление",
        "tags": [
          "catalogs"
        ]
      },
      "delete": {
        "operationId": "CatalogsController_deleteItDirection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Удалить ИТ-направление",
        "tags": [
          "catalogs"
        ]
      }
    },
...
  },
  "info": {
    "title": "CRM ИТ Школа РТК",
...
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "VendorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a3f0c2f0-4444-4a11-9a11-000000000030"
          },
          "name": {
            "type": "string",
            "example": "ООО «Базис»"
          },
          "contactInfo": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "contactName": {
            "type": "string",
            "nullable": true,
            "example": "Иванов Иван Иванович"
          },
          "contactPhone": {
            "type": "string",
            "nullable": true,
            "example": "+7 (900) 111-22-33"
          },
          "contactEmail": {
            "type": "string",
            "nullable": true,
            "example": "ivanov.ii@example.ru"
          },
          "contactChannel": {
            "type": "string",
            "nullable": true,
            "example": "Почта, Чат в ТГ"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-09-26T11:29:07.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-09-26T11:29:07.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateVendorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ООО «Базис»"
          },
          "contactInfo": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "contactName": {
            "type": "string",
            "nullable": true,
            "example": "Иванов Иван Иванович"
          },
          "contactPhone": {
            "type": "string",
            "nullable": true,
            "example": "+7 (900) 111-22-33"
          },
          "contactEmail": {
            "type": "string",
            "nullable": true,
            "example": "ivanov.ii@example.ru"
          },
          "contactChannel": {
            "type": "string",
            "nullable": true,
            "example": "Почта, Чат в ТГ"
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}